Corrigindo o método PUT para atualizar no banco

+++ src/main/java/teste_api/repository/CarteiraRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import teste_api.enitity.Carteira;
 
-public interface CarteiraRepository extends JpaRepository<Carteira, Integer> {
+public interface CarteiraRepository extends JpaRepository<Carteira, Long> {
+
 
 
 }
